Typedefining Qfloat as double rather than as float, to avoid warnings and possible errors in calculations.

+++ svm.cpp
@@ -7,7 +7,7 @@
 #include <stdarg.h>
 #include "svm.h"
 int libsvm_version = LIBSVM_VERSION;
-typedef float Qfloat;
+typedef double Qfloat;
 typedef signed char schar;
 #ifndef min
 template <class T> static inline T min(T x,T y) { return (x<y)?x:y; }
